[projectfirma/#1785] Allow Admins to Manage Attachment Types creating TaxonomyTrunkSimple and some code cleanup
