+++ Source/ProjectFirma.Web/Views/AttachmentRelationshipType/EditAttachmentRelationshipTypeViewData.cs
@@ -52,18 +52,4 @@ namespace ProjectFirma.Web.Views.AttachmentRelationshipType
             }
         }
     }
-
-    public class TaxonomyTrunkSimple
-    {
-        public int TaxonomyTrunkID { get; set; }
-        public string TaxonomyTrunkName { get; set; }
-        public string TaxonomyTrunkDescription { get; set; }
-
-        public TaxonomyTrunkSimple(ProjectFirmaModels.Models.TaxonomyTrunk taxonomyTrunk)
-        {
-            TaxonomyTrunkID = taxonomyTrunk.TaxonomyTrunkID;
-            TaxonomyTrunkName = taxonomyTrunk.TaxonomyTrunkName;
-            TaxonomyTrunkDescription = taxonomyTrunk.TaxonomyTrunkDescription;
-        }
-    }
 }
